Rename web/pages/os into web/pages/operatingsystems
Make the operatingsystem path more consistent and use the gsa name for
os.
