+++ gsa/src/web/pages/operatingsystems/component.js
