+++ gsa/src/web/pages/operatingsystems/dashboard/cvssdisplay.js
