+++ gsa/src/web/pages/operatingsystems/dashboard/index.js
