+++ gsa/src/web/pages/operatingsystems/dashboard/loaders.js
