+++ gsa/src/web/pages/operatingsystems/dashboard/severityclassdisplay.js
