+++ gsa/src/web/pages/operatingsystems/dashboard/vulnscoredisplay.js
