+++ gsa/src/web/pages/operatingsystems/detailspage.js
