+++ gsa/src/web/pages/operatingsystems/filterdialog.js
