+++ gsa/src/web/pages/operatingsystems/listpage.js
