+++ gsa/src/web/pages/operatingsystems/row.js
