+++ gsa/src/web/pages/operatingsystems/table.js
