+++ gsa/src/web/routes.js
@@ -67,8 +67,8 @@ import NotesPage from './pages/notes/listpage';
 import NoteDetailsPage from './pages/notes/detailspage';
 import NvtsPage from './pages/nvts/listpage';
 import NvtDetailsPage from './pages/nvts/detailspage';
-import OperatingSystemsPage from './pages/os/listpage';
-import OperatingSystemDetailsPage from './pages/os/detailspage';
+import OperatingSystemsPage from './pages/operatingsystems/listpage';
+import OperatingSystemDetailsPage from './pages/operatingsystems/detailspage';
 import OvaldefsPage from './pages/ovaldefs/listpage';
 import OvaldefDetailsPage from './pages/ovaldefs/detailspage';
 import OverridesPage from './pages/overrides/listpage';
